End with method in BookGlobalAPI name's getAllNewBook1
